factor out Diffraction2D
